requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
chroma-hnswlib==0.7.6
chromadb==0.6.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==3.3.1
gunicorn